Rename loki_fetch to loki_rpc
